Add quiet option for harmonisation operation
The harmonisation can now be done without any kind of output. Also the documentation was improved slightly.

The function now selects only the first element of a character vector with length > 1.

The usual warning for normalizePath() when non-existent paths are passed was suppressed in favour of a halting execution with an error message.