Fix a bug related to the 'serialno' column
The problem with the function in its current state lies in the fact that the data are pre-loaded by the preceding stack. Thus, regardless of the way data are loaded in the global environment, the data are reloaded from the table from scratch. Where the table does not have a 'serialno' column, the function fail. I guess the real problem is the fact that the function is tailor-made for a data table that comes with specific field format. This is something that could be reconsidered if it is to be used down the line. A note to this effect will be made in the documentation.